mq_timedreceive: 10-2.c: check timeout error rather than invalid parameter error
10-2.c now passes the invalid parameter to mq_timedreceive by specifying
a negative value to nanoseconds field, and checks whether EINVAL is not
returned when a message can be removed from the queue immediately.

However, this may not be a appropriate testing of mq_timedreceive.
POSIX spec of mq_timedreceive says:
    The validity of the abstime parameter need not be checked
    if a message can be removed from the message queue immediately.

This is open for interpretation of POSIX spec, and it was argued about in LKML.
    https://lkml.org/lkml/2012/3/14/589

The conclusion in LKML was that what POSIX spec says is "you don't *have*
to check validity". That is, it is allowed to check the validity.

On the other hand, actually, timeout error should not be returned when message
can be performed immediately.

Therefore, I modify the testcase to check timeout error rather than invalid
parameter error.

Signed-off-by: Akira Takeuchi <takeuchi.akr@jp.panasonic.com>
Signed-off-by: Kiyoshi Owada <owada.kiyoshi@jp.panasonic.com>
Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>