Fix fork/2-1.c
The allocated chunk corectness is asserted by casting it to double and
assigning a value. Later it's compared to the same value. The problem
with such approach is that in some cases the comparsion may be false,
due to implicit casts and nature of floating point operations.

So this patch removes the questionable floating point arithmetics and
instead of that fills allocated memory with pseudo random sequence which
is asserted later in the child.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>