numa_helper: fix compatibility with older libnuma
numa_helper relied on functions that do not exist
in old libnuma (<1), which caused that it couldn't be
compiled on old distros, for example RHEL 5.3.

Old libnuma and kernel also do not support MPOL_F_MEMS_ALLOWED,
in which case it will assume that any node with memory > 0 can
be used.

This patch adds compatibility for older versions, I tested it with:
numactl-0.9.8-7.el5
numactl-2.0.3-9.el6
numactl-2.0.7-6.el7

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Reviewed-by: Caspar Zhang <caspar@casparzhang.com>
Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>