mq_open/16-1: use tmp file to share info
In this test case, it uses a variable to share data between child and
parent processes. But after fork there is a copy of the variable in
child process and modify it will not affect the variable in the parent
process. Then when the child process call mq_open() before parent
process, the case will fail.

Use tmp file to replace the variable. Any modification in child process
can be seen in parent process.

Signed-off-by: Kang Kai <kai.kang@windriver.com>
Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>