bind/03, 21 rbind/38, 33:fix missing umount and error umount in these cases
The cases have the same problem that dirs are still in mount when their
targets are removed.for rbind/38,beside the problem above,there are two
error umount operations that umount dirs that are not in mount.
and all these result in making the test script trapped in a infinite loop.
To resolve these problems,all we need is just umount the mountpoints before
their targets are removed and delete the error umount operations.

Signed-off-by: Dou Xin <wq-doux@cn.fujitsu.com>
Signed-off-by: DAN LI <li.dan@cn.fujitsu.com>