build: Add INTERNAL_LIB option to lib.mk
We have a few internal only libraries in
order to share common testcases code. The
problem is that when creating library with:

...

LIB = libname

include $(top_srcdir)/include/mk/lib.mk

The LIB becomes MAKE_TARGET which later
becomes INSTALL_TARGET and gets installed.

This patch add INTERNAL_LIB variable to
lib.mk which is the same as LIB but the
MAKE_TARGET variable is not set on
'make install' i.e. when make target is
install.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>