kernel/security/securebits: Shuffle the ifdefs.
This shuffles the ifdefs around so that whole test
is ifdefed out when libcap2 is not installed.

This is done for two reasons. One is that the test
bails out when libcap2 wasn't present anyway and
the second is that the test still fails to compile
on older distros because of missing PR_GET_KEEPCAPS
in code that is not executed anyway (because libcap2
is not present there).

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>